feature: update logic with refresh token

+++ src/_security/repositories/security.repository.ts
@@ -53,12 +53,7 @@ export const securityRepository = {
     await securityCollection.deleteOne({deviceId, userId});
   },
   async deleteAllSessionExcludeCurrent (deviceId: string, userId: string){
-    return securityCollection.deleteMany({
-      $and: [
-        {deviceId: {$nin: [deviceId]}},
-        {userId: {$nin: [userId]}},
-      ]
-    })
+    return securityCollection.deleteMany({deviceId, userId})
   },
   async deleteCurrentSession(deviceId: string, userId: string){
     return securityCollection.deleteOne({deviceId, userId});
